Replace funflow-examples with examples from funflow2
