--- funflow-examples/ChangeLog.md
@@ -1,5 +1,0 @@
-# Revision history for funflow-examples
-
-## 0.1.0.0  -- YYYY-mm-dd
-
-* First version. Released on an unsuspecting world.
