--- funflow-examples/compile-and-run-c-files/README.md
@@ -1,4 +1,0 @@
-compile-and-run-c-files
-===
-
-This package defines an example workflow using Funflow.
