--- funflow-examples/funflow-examples.cabal
@@ -1,63 +1,0 @@
-name:                funflow-examples
-version:             0.1.0.0
-license:             MIT
-license-file:        LICENSE
-author:              Andreas Herrmann, Divesh Otwani
-maintainer:          andreas.herrmann@tweag.io
-build-type:          Simple
-extra-source-files:  ChangeLog.md
-cabal-version:       >=1.10
-
-executable external-c-computation
-  main-is:             Main.hs
-  build-depends:       base >=4.9 && < 5
-                     , cas-store
-                     , containers
-                     , funflow
-                     , path
-                     , path-io
-                     , text
-  default-language:    Haskell2010
-  hs-source-dirs:      compile-and-run-c-files
-
-executable makefile-tool
-  main-is:             Main.hs
-  other-modules:       Types
-                       Parse
-  build-depends:       base >=4.10 && < 5,
-                       cas-store,
-                       containers,
-                       funflow,
-                       path,
-                       path-io,
-                       text,
-                       katip,
-                       safe-exceptions,
-                       data-default,
-                       directory,
-                       parsec,
-                       unix,
-                       bytestring
-  hs-source-dirs:      makefile-tool/src
-  default-language:    Haskell2010
-
-executable omdb
-  main-is:             Main.hs
-  build-depends:       base >=4.9 && < 5
-                     , aeson
-                     , async
-                     , cas-hashable
-                     , cas-store
-                     , containers
-                     , data-default
-                     , funflow
-                     , lens
-                     , lens-aeson
-                     , optparse-generic
-                     , path
-                     , path-io
-                     , safe-exceptions
-                     , text
-                     , wreq
-  default-language:    Haskell2010
-  hs-source-dirs:      omdb
