+++ funflow-examples/makefile-tool/src/Types.hs
@@ -13,6 +13,7 @@ type Set = Set.Set
 type SourceFile = String     -- | Name of source file
 type TargetFile = String     -- | Name of target file
 type BuildFile  = String     -- | Either source xor target
+type Command    = String
 
 data MakeFile where
   MakeFile :: { sourceFiles :: Set SourceFile
@@ -25,17 +26,10 @@ data MakeRule where
   MakeRule :: TargetFile -> Set BuildFile -> Command -> MakeRule
   deriving (Eq, Ord, Show)
 
-type Command = String
-
 -- Makefile error type
 newtype MFError = MFError String
 
-
-
 -- | Accessors
 --------------------------------------------------------------------------------
-
 mkRuleTarNm :: MakeRule -> String
 mkRuleTarNm (MakeRule tf _ _) = tf
-
-
