--- funflow2/funflow-examples/makefile-tool/test/factorial.cpp
@@ -1,8 +1,0 @@
-#include "functions.h"
-
-int factorial(int n){
-    if(n!=1){
-	return(n * factorial(n-1));
-    }
-    else return 1;
-}
