--- funflow2/funflow-examples/makefile-tool/test/functions.h
@@ -1,2 +1,0 @@
-void print_hello();
-int factorial(int n);
