--- funflow2/funflow-examples/makefile-tool/test/hello.cpp
@@ -1,8 +1,0 @@
-#include <iostream>
-#include "functions.h"
-
-using std::cout;
-
-void print_hello(){
-   cout << "Hello World!";
-}
