--- funflow2/funflow-examples/makefile-tool/test/main.cpp
@@ -1,13 +1,0 @@
-#include <iostream>
-#include "functions.h"
-
-using std::cout;
-using std::endl;
-
-
-int main(){
-    print_hello();
-    cout << endl;
-    cout << "The factorial of 5 is " << factorial(5) << endl;
-    return 0;
-}
